Moved IngestConfigurator interface to ingest package
