+++ Core/src/org/sleuthkit/autopsy/casemodule/AddImageAction.java
@@ -19,6 +19,7 @@
 
 package org.sleuthkit.autopsy.casemodule;
 
+import org.sleuthkit.autopsy.ingest.IngestConfigurator;
 import java.awt.Component;
 import java.awt.Dialog;
 import java.awt.event.ActionEvent;
