+++ Core/src/org/sleuthkit/autopsy/ingest/IngestConfigurator.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.sleuthkit.autopsy.casemodule;
+package org.sleuthkit.autopsy.ingest;
 
 import java.util.List;
 import javax.swing.JPanel;
